Move Plugins/Digitization to Core/Digitization
