+++ Core/src/Digitization/CMakeLists.txt
@@ -2,4 +2,6 @@ target_sources(
   ActsCore
   PRIVATE
     CartesianSegmentation.cpp
+    DigitizationModule.cpp
+    PlanarModuleStepper.cpp
 )
