+++ Core/src/Digitization/PlanarModuleStepper.cpp
@@ -10,10 +10,10 @@
 // PlanarModuleStepper.cpp, Acts project
 ///////////////////////////////////////////////////////////////////
 
-#include "Acts/Plugins/Digitization/PlanarModuleStepper.hpp"
+#include "Acts/Digitization/PlanarModuleStepper.hpp"
 
 #include "Acts/Definitions/Algebra.hpp"
-#include "Acts/Plugins/Digitization/DigitizationModule.hpp"
+#include "Acts/Digitization/DigitizationModule.hpp"
 #include "Acts/Surfaces/Surface.hpp"
 #include "Acts/Utilities/Intersection.hpp"
