+++ Examples/Algorithms/Printers/CMakeLists.txt
@@ -8,7 +8,7 @@ target_include_directories(
   PUBLIC $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}>)
 target_link_libraries(
   ActsExamplesPrinters
-  PUBLIC ActsCore ActsFatras ActsPluginDigitization ActsExamplesFramework)
+  PUBLIC ActsCore ActsFatras ActsExamplesFramework)
 
 install(
   TARGETS ActsExamplesPrinters
