+++ Examples/Io/Csv/src/CsvPlanarClusterReader.cpp
@@ -9,7 +9,7 @@
 #include "ActsExamples/Io/Csv/CsvPlanarClusterReader.hpp"
 
 #include "Acts/Definitions/Units.hpp"
-#include "Acts/Plugins/Digitization/PlanarModuleCluster.hpp"
+#include "Acts/Digitization/PlanarModuleCluster.hpp"
 #include "Acts/Plugins/Identification/IdentifiedDetectorElement.hpp"
 #include "Acts/Surfaces/Surface.hpp"
 #include "ActsExamples/EventData/GeometryContainers.hpp"
