+++ Examples/Python/src/Digitization.cpp
@@ -6,7 +6,7 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-#include "Acts/Plugins/Digitization/PlanarModuleStepper.hpp"
+#include "Acts/Digitization/PlanarModuleStepper.hpp"
 #include "Acts/Plugins/Python/Utilities.hpp"
 #include "ActsExamples/Digitization/DigitizationAlgorithm.hpp"
 #include "ActsExamples/Digitization/DigitizationConfig.hpp"
