+++ Tests/DownstreamProject/CMakeLists.txt
@@ -10,7 +10,6 @@ find_package(
     Fatras
     PluginAutodiff
     PluginDD4hep
-    PluginDigitization
     PluginIdentification
     PluginJson
     PluginLegacy
@@ -30,7 +29,6 @@ target_link_libraries(
     ActsCore
     ActsFatras
     ActsPluginDD4hep
-    ActsPluginDigitization
     ActsPluginIdentification
     ActsPluginJson
     ActsPluginLegacy
