+++ Tests/UnitTests/Core/SpacePointFormation/CMakeLists.txt
@@ -1,4 +1,2 @@
-set(unittest_extra_libraries ActsPluginDigitization)
-
 add_unittest(SingleHitSpacePointBuilder SingleHitSpacePointBuilderTests.cpp)
 add_unittest(DoubleHitSpacePointBuilder DoubleHitSpacePointBuilderTests.cpp)
