+++ Tests/UnitTests/Core/SpacePointFormation/DoubleHitSpacePointBuilderTests.cpp
@@ -11,7 +11,7 @@
 
 #include "Acts//Digitization/DoubleHitSpacePointBuilder.hpp"
 #include "Acts/Definitions/Algebra.hpp"
-#include "Acts/Plugins/Digitization/PlanarModuleCluster.hpp"
+#include "Acts/Digitization/PlanarModuleCluster.hpp"
 #include "Acts/Surfaces/PlaneSurface.hpp"
 #include "Acts/Surfaces/RectangleBounds.hpp"
 #include "Acts/Tests/CommonHelpers/DetectorElementStub.hpp"
