--- Tests/UnitTests/Plugins/Digitization/CMakeLists.txt
@@ -1,3 +1,0 @@
-set(unittest_extra_libraries ActsPluginDigitization)
-
-add_unittest(PlanarModuleStepper PlanarModuleStepperTests.cpp)
